% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{custom_bar_plot}
\alias{custom_bar_plot}
\title{Create custom bar plot}
\usage{
custom_bar_plot(
  data,
  x,
  y,
  fill,
  xlab,
  ylab,
  title,
  ylim = NULL,
  ybreaks = waiver(),
  fill_name,
  fill_values,
  fill_labels = NULL,
  show_legend = TRUE,
  legend_position = NULL,
  facets = FALSE,
  facet_var = NULL,
  facet_scales = NULL,
  errorbar = FALSE,
  ymin = NULL,
  ymax = NULL,
  nutrients = FALSE,
  rw = NULL,
  rw_labels = NULL,
  text_group = NULL
)
}
\description{
Create custom bar plot
}
