% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fulcrum_table_try}
\alias{fulcrum_table_try}
\title{Modified version of fulcrumr::fulcrum_table}
\usage{
fulcrum_table_try(table_name)
}
\arguments{
\item{table_name}{Name of Fulcrum table}
}
\description{
Catch connection errors and return NULL
}
