% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{get_bmi_samples}
\alias{get_bmi_samples}
\title{Get BMI samples table}
\usage{
get_bmi_samples(yrs = c("2022", "2023", "2024"))
}
\arguments{
\item{yrs}{Vector of four-digit years as a character}
}
\description{
Get BMI samples table
}
