% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{get_bmi_site}
\alias{get_bmi_site}
\title{Get BMI site visit table}
\usage{
get_bmi_site(yrs = c("2023", "2024"))
}
\arguments{
\item{yrs}{Vector of four-digit years as a character}
}
\description{
Get BMI site visit table
}
