% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_gbl}
\alias{get_gbl}
\title{Get group B location (A, B, C, N/A) from group B name}
\usage{
get_gbl(group_b_name)
}
\arguments{
\item{group_b_name}{Vector of group B names, e.g., 10DASHa2023}
}
\description{
Get group B location (A, B, C, N/A) from group B name
}
