% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrients.R
\name{get_gbm}
\alias{get_gbm}
\title{Get group B method from group B name}
\usage{
get_gbm(data)
}
\arguments{
\item{data}{Dataframe with group_b_name column}
}
\description{
Get group B method from group B name
}
