% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrophytes.R
\name{get_mac}
\alias{get_mac}
\title{Get main macrophyte table}
\usage{
get_mac(yr = c("2022", "2023", "2024"))
}
\arguments{
\item{yr}{Four-digit year as a character}
}
\description{
Get main macrophyte table
}
