% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrophytes.R
\name{get_mac_samples}
\alias{get_mac_samples}
\title{Get macrophyte samples table}
\usage{
get_mac_samples(yr)
}
\arguments{
\item{yr}{Four-digit year as a character}
}
\description{
Get macrophyte samples table
}
