% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrophytes.R
\name{get_mac_site}
\alias{get_mac_site}
\title{Get macrophyte site visit table}
\usage{
get_mac_site(yr, mac)
}
\arguments{
\item{yr}{Four-digit year as a character}

\item{mac}{Macrophyte dataframe returned from get_mac}
}
\description{
Get macrophyte site visit table
}
