% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrients.R
\name{get_nut_samples}
\alias{get_nut_samples}
\title{Get nutrient samples table}
\usage{
get_nut_samples(yr)
}
\arguments{
\item{yr}{Four-digit year as a character}
}
\description{
Get nutrient samples table
}
