% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrophytes.R
\name{isobiweek}
\alias{isobiweek}
\title{Modified version of lubridate::isoweek}
\usage{
isobiweek(x, start_date)
}
\arguments{
\item{x}{Date-time object}

\item{start-date}{Date of Monday in first sampling week}
}
\description{
The goal is to assign a date to a biweek such that the first biweek in a period
has the same date as the Monday of the week when sampling started.
}
