% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mean_se}
\alias{mean_se}
\title{Calculate mean and standard error}
\usage{
mean_se(grp_df, column_name)
}
\arguments{
\item{grp_df}{Grouped dataframe}

\item{column_name}{Name of column in grp_df to calculate mean and SE}
}
\description{
Calculate mean and standard error
}
