% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{prep_bmi}
\alias{prep_bmi}
\title{Prepare main BMI tables}
\usage{
prep_bmi(yrs = c("2022", "2023", "2024"))
}
\arguments{
\item{yrs}{Vector of four-digit years as a character}
}
\description{
Prepare main BMI tables
}
