% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{prep_bmi_samples}
\alias{prep_bmi_samples}
\title{Prepare BMI samples table}
\usage{
prep_bmi_samples(yrs = c("2022", "2023", "2024"))
}
\arguments{
\item{yrs}{Vector of four-digit years as a character}
}
\description{
Prepare BMI samples table
}
