% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{prep_bmi_site}
\alias{prep_bmi_site}
\title{Prepare BMI site visit table}
\usage{
prep_bmi_site(yrs = c("2023", "2024"))
}
\arguments{
\item{yrs}{Vector of four-digit years as a character}
}
\description{
Prepare BMI site visit table
}
