% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{prep_bmi_spatial}
\alias{prep_bmi_spatial}
\title{Prepare BMI spatial table}
\usage{
prep_bmi_spatial(bmi_samples)
}
\arguments{
\item{bmi_samples}{BMI samples dataframe returned from prep_bmi_samples}
}
\description{
Prepare BMI spatial table
}
