% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrophytes.R
\name{prep_mac}
\alias{prep_mac}
\title{Prepare main macrophyte table}
\usage{
prep_mac(yr = c("2022", "2023", "2024"))
}
\arguments{
\item{yr}{Four-digit year as a character}
}
\description{
Prepare main macrophyte table
}
