% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrophytes.R
\name{prep_mac_freq}
\alias{prep_mac_freq}
\title{Prepare macrophyte frequency table}
\usage{
prep_mac_freq(mac_species)
}
\arguments{
\item{mac_species}{Macrophyte species dataframe returned from prep_mac_species}
}
\description{
Prepare macrophyte frequency table
}
