% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrophytes.R
\name{prep_mac_samples}
\alias{prep_mac_samples}
\title{Prepare macrophyte samples table}
\usage{
prep_mac_samples(yr, mac_site)
}
\arguments{
\item{yr}{Four-digit year as a character}

\item{mac_site}{Macrophyte site visit dataframe returned from get_mac_site}
}
\description{
Prepare macrophyte samples table
}
