% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrophytes.R
\name{prep_mac_site}
\alias{prep_mac_site}
\title{Prepare macrophyte site visit table}
\usage{
prep_mac_site(yr, mac)
}
\arguments{
\item{yr}{Four-digit year as a character}

\item{mac}{Macrophyte dataframe returned from get_mac}
}
\description{
Prepare macrophyte site visit table
}
