% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrophytes.R
\name{prep_mac_species}
\alias{prep_mac_species}
\title{Prepare macrophyte species table}
\usage{
prep_mac_species(mac_samples, rakes, species_df = TahoeKeysCMT::species_df)
}
\arguments{
\item{mac_samples}{Macrophyte samples dataframe returned from get_mac_samples}

\item{rakes}{Macrophyte rakes dataframe returned from prep_rakes}

\item{species_df}{Species dataframe included with this package}
}
\description{
Prepare macrophyte species table
}
