% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrients.R
\name{prep_nut}
\alias{prep_nut}
\title{Prepare main nutrient table}
\usage{
prep_nut(yr = c("2022", "2023", "2024"))
}
\arguments{
\item{yr}{Four-digit year as a character}
}
\description{
Prepare main nutrient table
}
