% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrients.R
\name{prep_nut_samples}
\alias{prep_nut_samples}
\title{Prepare nutrient samples table}
\usage{
prep_nut_samples(yr, nut_site)
}
\arguments{
\item{yr}{Four-digit year as a character}

\item{nut_site}{Nutrient site visit dataframe returned from get_nut_site}
}
\description{
Prepare nutrient samples table
}
