% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrients.R
\name{prep_nut_site}
\alias{prep_nut_site}
\title{Prepare nutrient site visit table}
\usage{
prep_nut_site(yr, nut)
}
\arguments{
\item{yr}{Four-digit year as a character}

\item{nut}{Nutrient dataframe returned from get_nut}
}
\description{
Prepare nutrient site visit table
}
