% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrients.R
\name{prep_nut_spatial}
\alias{prep_nut_spatial}
\title{Prepare nutrient spatial table}
\usage{
prep_nut_spatial(nut_samples)
}
\arguments{
\item{nut_samples}{Nutrient samples dataframe returned from prep_nut_samples}
}
\description{
Prepare nutrient spatial table
}
