% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrophytes.R
\name{prep_rakes}
\alias{prep_rakes}
\title{Prepare macrophyte rakes table}
\usage{
prep_rakes(yr, mac_site, mac_samples)
}
\arguments{
\item{yr}{Four-digit year as a character}

\item{mac_site}{Macrophyte site visit dataframe returned from get_mac_site}

\item{mac_samples}{Macrophyte samples dataframe returned from get_mac_samples}
}
\description{
Prepare macrophyte rakes table
}
