% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrophytes.R
\name{prep_rakes_spatial}
\alias{prep_rakes_spatial}
\title{Prepare rakes spatial table}
\usage{
prep_rakes_spatial(yr, mac_species, rakes)
}
\arguments{
\item{yr}{Four-digit year as a character}

\item{mac_species}{Macrophyte species dataframe returned from prep_mac_species}

\item{rakes}{Macrophyte rakes dataframe returned from prep_rakes}
}
\description{
Prepare rakes spatial table
}
