% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrophytes.R
\name{prepare_mac_samples}
\alias{prepare_mac_samples}
\title{Prrepare macrophyte samples table}
\usage{
prepare_mac_samples(yr, mac_site)
}
\arguments{
\item{yr}{Four-digit year as a character}

\item{mac_site}{Macrophyte site visit dataframe returned from get_mac_site}
}
\description{
Prrepare macrophyte samples table
}
