% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{process_photo_id}
\alias{process_photo_id}
\title{Process Fulcrum photo ID string}
\usage{
process_photo_id(x)
}
\arguments{
\item{x}{Vector of strings potentially comprised of more than one photo ID}
}
\description{
Process Fulcrum photo ID string
}
