% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrophytes.R
\name{process_samples}
\alias{process_samples}
\title{Process macrophyte samples for relative abundance and plant health}
\usage{
process_samples(mac_samples, type = c("relative_abundance", "rating_score"))
}
\arguments{
\item{mac_samples}{Macrophyte samples dataframe returned from get_mac_samples}

\item{type}{Type of response: relative_abundance or rating_score}
}
\description{
Process macrophyte samples for relative abundance and plant health
}
