% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrophytes.R
\name{get_mac_samples}
\alias{get_mac_samples}
\title{Get macrophyte samples table}
\usage{
get_mac_samples(yr, mac_site)
}
\arguments{
\item{yr}{Four-digit year as a character}

\item{mac_site}{Macrophyte site visit dataframe returned from get_mac_site}
}
\description{
Get macrophyte samples table
}
